Move files for better separation of exploration highlighting
