+++ applications/conceptual-model-editor/src/catalog/exploration-highlighting/highlighting-catalog-styles.css
