+++ applications/conceptual-model-editor/src/context/highlighting-exploration-mode.tsx
