+++ applications/conceptual-model-editor/src/diagram/features/highlighting/exploration/shared-style/exploration-highlighting-styles.css
